Moved css util functions in ActionList.tsx to their own file
